refactor(assets): remove REFERENCE render-mode (keep bundle helpers)
Removes AssetMode.REFERENCE, LoadedAssets, PJX_ASSETS_HEADER, should_emit_reference_url,
and all REFERENCE branches from assets.py/renderer.py/client.py/base.py. AssetPolicy now
only carries js_mode/css_mode. Bundle helpers (Finder, asset_manifest, resolver_with_hash,
hashed_filename, etc.) are untouched. Rescued non-REFERENCE tests into css_collection,
all_assets, and javascript_collection before deleting test_asset_reference_mode.py.
